pyproject.toml
[project]
name = "hexswitch"
version = "0.1.0"
description = "Hexagonal runtime switchboard for config-driven microservices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "websockets>=12.0",
    # OpenTelemetry
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-semantic-conventions>=0.60b0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    # FastAPI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    # Pydantic
    "pydantic>=2.5.0",
]
authors = [
    {name = "HexSwitch Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
hexswitch = "hexswitch.app:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.1.0",
    "mypy>=1.5.0",
    "radon>=6.0.0",
    "typer>=0.9.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "devtool @ file:./devops/devtool",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 200
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py are common

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=src/hexswitch",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "docker: Docker integration tests",
    "timeout: Test timeout marker (requires pytest-timeout)",
    "fast: Fast tests (< 1s)",
    "security: Security-focused tests",
    "stress: Stress and robustness tests",
    "slow: Slow tests (< 60s)",
    "edge_cases: Edge case tests",
]

[tool.coverage.run]
source = ["src/hexswitch"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 50

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.radon]
exclude = [
    "tests/*",
    "*/test_*.py",
]

